+++ app/Http/Controllers/RulesOfProceduresController.php
@@ -7,6 +7,9 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
+use App\Facades\CustomLog;
+use App\Http\Requests\StoreRulesOfProcedureRequest;
+use App\Http\Requests\UpdateRulesOfProcedureRequest;
 
 class RulesOfProceduresController extends Controller
 {
@@ -21,8 +24,8 @@ class RulesOfProceduresController extends Controller
         if($standardId == null){
             return redirect('/');
         }
-        $documents = Document::where('doc_category', 'rules_procedure')->where([['standard_id', $standardId],['team_id',Auth::user()->current_team_id]])->get();
-        $folder = "rules_of_procedure";
+        $documents = Document::where([ ['doc_category', 'rules_procedure'], ['standard_id', $standardId], ['team_id',Auth::user()->current_team_id] ])->get();
+        $folder = \Str::snake($this::getCompanyName())."/rules_of_procedure";
         $route_name = "rules-of-procedures";
         return view('documents.index', compact('documents', 'folder', 'route_name'));
     }
@@ -34,7 +37,8 @@ class RulesOfProceduresController extends Controller
      */
     public function create()
     {
-        return view('documents.create', ['url'=> route('rules-of-procedures.store'), 'back' => route('rules-of-procedures.index')]);
+        $this->authorize('create', Document::class);
+        return view('documents.create', ['url' => route('rules-of-procedures.store'), 'back' => route('rules-of-procedures.index')]);
     }
 
     /**
@@ -43,40 +47,29 @@ class RulesOfProceduresController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(StoreRulesOfProcedureRequest $request)
     {
+        $this->authorize('create', Document::class);
         $document = new Document();
-
-        $messages = array(
-            'file.required' => 'Izaberite fajl',
-            'file.mimes' => 'Fajl mora biti u PDF formatu',
-            'document_name.required' => 'Unesite naziv dokumenta',
-            'document_name.max' => 'Naziv dokumenta ne sme biti duži od 255 karaktera',
-            'document_name.unique' => 'Već postoji dokument sa takvim nazivom',
-            'document_version.required' => 'Unesite verziju dokumenta'
-        );
-
-        $validatedData = $request->validate([
-            'document_name' => 'required|unique:documents|max:255',
-            'document_version' => 'required',
-            'file' => 'required|max:10000|mimes:pdf'
-        ], $messages);
 
         $document->doc_category = 'rules_procedure';
         $document->document_name = $request->document_name;
         $document->version = $request->document_version;
+        $document->team_id = Auth::user()->current_team_id;
+        $document->user_id = Auth::user()->id;
+
+        $upload_path = "/public/".\Str::snake($this::getCompanyName())."/rules_of_procedure";
 
         if($request->file('file')){
             $file = $request->file;
-
             $name = $file->getClientOriginalName();
             $document->file_name = 'rules_of_procedure_'.time().'.'.$file->getClientOriginalExtension();
             $standard = $this::getStandard();
             $document->standard_id = $standard;
             $document->save();
-            $file->storeAs('/public/rules_of_procedure', $document->file_name);
-
+            $file->storeAs($upload_path, $document->file_name);
             $request->session()->flash('status', 'Dokument je uspešno sačuvan!');
+            CustomLog::info('Dokument Poslovnik "'.$document->document_name.'" kreiran. Korisnik: '.\Auth::user()->name.', '.\Auth::user()->email.', '.date('d.m.Y').' u '.date('H:i:s'), 'Firma-'.\Auth::user()->current_team_id);
             return redirect('/rules-of-procedures');
         }
     }
@@ -100,6 +93,7 @@ class RulesOfProceduresController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize('update', Document::find($id));
         $document = Document::findOrFail($id);
         return view('documents.edit', ['document' => $document, 'url' => route('rules-of-procedures.update', $document->id), 'folder' => 'rules_of_procedure', 'back' => route('rules-of-procedures.index')]);
     }
@@ -111,26 +105,16 @@ class RulesOfProceduresController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(UpdateRulesOfProcedureRequest $request, $id)
     {
+        $this->authorize('update', Document::find($id));
         $document = Document::findOrFail($id);
-
-        $messages = array(
-            'file.mimes' => 'Fajl mora biti u PDF formatu',
-            'document_name.required' => 'Unesite naziv dokumenta',
-            'document_name.max' => 'Naziv dokumenta ne sme biti duži od 255 karaktera',
-            'document_version.required' => 'Unesite verziju dokumenta'
-        );
-
-        $validatedData = $request->validate([
-            'document_name' => 'required|max:255',
-            'document_version' => 'required',
-            'file' => 'max:10000|mimes:pdf'
-        ], $messages);
 
         $document->doc_category = 'rules_procedure';
         $document->document_name = $request->document_name;
         $document->version = $request->document_version;
+
+        $upload_path = "/public/".\Str::snake($this::getCompanyName())."/rules_of_procedure";
 
         if($request->file('file')){
             $file = $request->file;
@@ -138,11 +122,12 @@ class RulesOfProceduresController extends Controller
             $document->file_name = 'rules_of_procedure_'.time().'.'.$file->getClientOriginalExtension();
             $standard = $this::getStandard();
             $document->standard_id = $standard;
-            $file->storeAs('/public/rules_of_procedure', $document->file_name);
+            $file->storeAs($upload_path, $document->file_name);
         }
 
         $document->save();
         $request->session()->flash('status', 'Dokument je uspešno izmenjen!');
+        CustomLog::info('Dokument Poslovnik "'.$document->document_name.'" izmenjen. Korisnik: '.\Auth::user()->name.', '.\Auth::user()->email.', '.date('d.m.Y').' u '.date('H:i:s'), 'Firma-'.\Auth::user()->current_team_id);
         return redirect('/rules-of-procedures');
     }
 
@@ -154,8 +139,11 @@ class RulesOfProceduresController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('delete', Document::find($id));
+        $doc_name = Document::find($id)->document_name;
+        
         if(Document::destroy($id)){
-            //logic for deleting the file from the server
+            CustomLog::info('Dokument Poslovnik "'.$doc_name.'" uklonjen. Korisnik: '.\Auth::user()->name.', '.\Auth::user()->email.', '.date('d.m.Y').' u '.date('H:i:s'), 'Firma-'.\Auth::user()->current_team_id);
             return back()->with('status', 'Dokument je uspešno uklonjen');
         }else{
             return back()->with('status', 'Došlo je do greške! Pokušajte ponovo.');
